+++ src/Volo/FrontendBundle/Twig/VendorExtension.php
@@ -2,6 +2,7 @@
 
 namespace Volo\FrontendBundle\Twig;
 
+use Volo\FrontendBundle\Service\EventService;
 use Volo\FrontendBundle\Service\ScheduleService;
 
 class VendorExtension extends \Twig_Extension
@@ -12,11 +13,18 @@ class VendorExtension extends \Twig_Extension
     protected $scheduleService;
 
     /**
-     * @param ScheduleService $scheduleService
+     * @var EventService
      */
-    public function __construct(ScheduleService $scheduleService)
+    protected $eventService;
+
+    /**
+     * @param ScheduleService $scheduleService
+     * @param EventService $eventService
+     */
+    public function __construct(ScheduleService $scheduleService, EventService $eventService)
     {
         $this->scheduleService = $scheduleService;
+        $this->eventService = $eventService;
     }
 
     /**
@@ -26,6 +34,7 @@ class VendorExtension extends \Twig_Extension
     {
         return array(
             new \Twig_SimpleFunction('isVendorOpen', array($this->scheduleService, 'isVendorOpen')),
+            new \Twig_SimpleFunction('isVendorFloodFeatureClosed', array($this->eventService, 'isVendorFloodFeatureClosed')),
             new \Twig_SimpleFunction('getDailySchedules', array($this->scheduleService, 'getDailySchedules')),
             new \Twig_SimpleFunction('getNextDayPeriods', array($this->scheduleService, 'getNextDayPeriods')),
             new \Twig_SimpleFunction('getTimePickerJsonValues', array($this->scheduleService, 'getTimePickerJsonValues')),